+++ CHANGELOG.md
@@ -4,8 +4,9 @@ This changelog covers all three packages, as they are (for now) updated as a who
 
 ## Unreleased
 
+- `@tomic/react` can now be used without `@tomic/lib` - it re-exports the library
 - Allow `@tomic/lib` to be used in non-browser (Node) context #128
-- Add `useMarkdown` function to `@tomic/react`.
+- Add `useMarkdown` function to `@tomic/react`
 - Make search result previews smaller
 - Fetch full collections when showing CollectionCard
 - `useResource` defaults to not accepting incomplete resources